+++ resources/views/tasks/index.blade.php
@@ -13,13 +13,13 @@
                 </x-primary-button>
         {{ html()->form()->close()}}
         
-        @auth
+        @can('create', App\Models\Task::class)
         <div>
             <x-primary-link href="{{ route('tasks.create') }}">
                 {{ __('Create task') }}
             </x-primary-link>
         </div>
-        @endauth
+        @endcan
 
     </div>
 
@@ -32,9 +32,9 @@
                 <th>{{ __('Author') }}</th>
                 <th>{{ __('Executor') }}</th>
                 <th>{{ __('Created at') }}</th>
-                @auth
+                @can('update', $tasks[0] ?? null)
                     <th>{{ __('Actions')}} </th>
-                @endauth
+                @endcan
             </tr>
         </thead>
         <tbody>
@@ -51,7 +51,7 @@
                     <td>{{  $task->author->name }}</td>
                     <td>{{  $task->assignedTo?->name }}</td>
                     <td>{{  $task->created_at->format('d.m.Y') }}</td>                        
-                    @auth
+                    @can('update', $task)
                         <td>
                             @can('delete', $task)
                                 <a data-confirm="Вы уверены?" data-method="delete" class="text-red-600 hover:text-red-900" href="{{ route('tasks.destroy', $task->id)  }}">
@@ -62,7 +62,7 @@
                                 {{ __('Change') }}
                             </a>
                         </td>
-                    @endauth
+                    @endcan
                 </tr>
             @endforeach
         </tbody>
